requirements.txt
playsound==1.3.0
pillow==11.1.0
pyinstaller==6.12.0